Re-arrange models, specify db_table
