+++ opencivicdata/legislative/models/__init__.py
@@ -7,8 +7,3 @@ from .vote import (VoteEvent, VoteCount, PersonVote, VoteSource)
 from .event import (Event, EventLocation, EventMedia, EventMediaLink, EventDocument, EventLink,
                     EventSource, EventParticipant, EventAgendaItem, EventRelatedEntity,
                     EventAgendaMedia, EventAgendaMediaLink, EventDocumentLink)
-from .elections import (Election, ElectionIdentifier, Candidacy, CandidacySource, Party)
-from .elections.contests import (ContestBase, ContestIdentifier, ContestSource, 
-                                 BallotMeasureContest, BallotMeasureContestOption, 
-                                 CandidateContest, CandidateContestPost, PartyContest,
-                                 PartyContestOption, RetentionContest)
